+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added property to Cisco Codec VideoComponent to determine if the codec supports video mute based on system software version.
 
 ## [17.2.0] - 2021-01-14
 ### Added